+++ HISTORY.rst
@@ -11,6 +11,9 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
   < 3.4.
 - #636: [Linux] *connections functions may swallow errors and return an
   incomplete list of connnections.
+- #637: [UNIX] raise exception if trying to send signal to Process PID 0 as it
+  will affect os.getpid()'s process group instead of PID 0.
+
 
 3.0.0 - 2015-06-13
 ==================